Move some properties to Atron class
