+++ units/ally/atron.py
@@ -1,12 +1,19 @@
 """Abstract base class for all atron units."""
 
-from utils import Counter
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from utils import DEFAULT_CONTEXT, Context, Counter
+
+if TYPE_CHECKING:
+    from utils import Icon
 
 
 class Atron:
     """Abstract base class for all atron units."""
 
-    def __init__(self, health: int) -> None:
+    def __init__(self, health: int, capacity: int = 0) -> None:
         """Initialize a atron unit.
 
         The atron's health must be at least 1.
@@ -16,10 +23,13 @@ class Atron:
 
         Args:
             health (int): The atron's maximum health.
+            capacity (int, optional): The atron's maximum mineral capacity.
         """
         if health <= 0:
             raise ValueError("Atron health must be 1 or greater")
         self._health = Counter(value=health, max_value=health)
+        self._payload = Counter(value=0, max_value=capacity)
+        self._context: Context = DEFAULT_CONTEXT
 
     @property
     def health(self) -> Counter:
@@ -29,6 +39,38 @@ class Atron:
             Counter: The current health.
         """
         return self._health
+
+    @property
+    def payload(self) -> Counter:
+        """The atron's mineral payload.
+
+        Returns:
+            Counter: The mineral payload.
+        """
+        return self._payload
+
+    @property
+    def context(self) -> Context:
+        """The context surrounding this atron.
+
+        Returns:
+            Context: The context.
+        """
+        return self._context
+
+    @context.setter
+    def context(self, new_context: Context) -> None:
+        """Set the context surrounding this atron.
+
+        Args:
+            new_context (Context): The new context.
+        """
+        self._context = new_context
+
+    @property
+    def icon(self) -> Icon:
+        """The icon of this atron."""
+        raise NotImplementedError("Drone subtypes must implement icon")
 
     def __str__(self):
         """Return the string representation of this object.
